Dataset-tck: Move helper classes for dataset_condenser under dataset.tck
Also move dataset_tck module under tck as io_utils.
