+++ src/dataset/tck/constants.py
@@ -1,0 +1,2 @@
+
+SRCFILE_KEY = 'source_file_acquisition_full'
